+++ example/serialize_cyclical_list.g.dart
@@ -88,14 +88,15 @@ abstract class _$CourseSerializable extends SerializableMap {
 
 _Student__Constructor([positionalParams, namedParams]) => new Student();
 
-const $$Student_fields_id = const DeclarationMirror(type: int);
-const $$Student_fields_name = const DeclarationMirror(type: String);
+const $$Student_fields_id = const DeclarationMirror(name: 'id', type: int);
+const $$Student_fields_name =
+    const DeclarationMirror(name: 'name', type: String);
 const $$Student_fields_courses =
-    const DeclarationMirror(type: const [List, Course]);
+    const DeclarationMirror(name: 'courses', type: const [List, Course]);
 
 const StudentClassMirror =
     const ClassMirror(name: 'Student', constructors: const {
-  '': const FunctionMirror($call: _Student__Constructor)
+  '': const FunctionMirror(name: '', $call: _Student__Constructor)
 }, annotations: const [
   cyclical
 ], fields: const {
@@ -113,14 +114,15 @@ const StudentClassMirror =
 ]);
 _Course__Constructor([positionalParams, namedParams]) => new Course();
 
-const $$Course_fields_id = const DeclarationMirror(type: int);
-const $$Course_fields_beginDate = const DeclarationMirror(type: DateTime);
+const $$Course_fields_id = const DeclarationMirror(name: 'id', type: int);
+const $$Course_fields_beginDate =
+    const DeclarationMirror(name: 'beginDate', type: DateTime);
 const $$Course_fields_students =
-    const DeclarationMirror(type: const [List, Student]);
+    const DeclarationMirror(name: 'students', type: const [List, Student]);
 
 const CourseClassMirror =
     const ClassMirror(name: 'Course', constructors: const {
-  '': const FunctionMirror($call: _Course__Constructor)
+  '': const FunctionMirror(name: '', $call: _Course__Constructor)
 }, annotations: const [
   cyclical
 ], fields: const {