+++ lib/action.dart
@@ -4,7 +4,7 @@ import 'package:built_mirrors/init_mirrors_generator.dart';
 import 'package:dson/generator.dart';
 import 'package:source_gen/source_gen.dart';
 
-/// Creates a `built_mirrors` generator-builder phase.
+/// Creates a `built_mirrors` generator-builder action.
 ///
 /// Example: {@example /tool/build.dart}
 /// Example: {@example /tool/watch.dart}