+++ lib/src/serializer.dart
@@ -148,17 +148,6 @@ Object _serializeObject(obj, depth, exclude, fieldName) {
   return result;
 }
 
-Map _uIdFromClassCache = {};
-
-String _getUIdAttrFromClass(ClassMirror cm) =>
-    (_uIdFromClassCache
-      ..putIfAbsent(cm, () =>
-        cm.fields.values
-            .firstWhere((v) => v.annotations?.contains((a) => a == uId) ?? false, orElse: () => null)
-            ?.name ?? 'id'
-      )
-    )[cm];
-
 /// Checks the DeclarationMirror [variable] for annotations and adds
 /// the value to the [result] map. If there's no [SerializedName] annotation
 /// with a different name set it will use the name of [symbol].
@@ -170,20 +159,13 @@ void _pushField(String fieldName, DeclarationMirror variable, SerializableMap ob
   Object value = obj[fieldName];
 //  _serLog.finer("Start serializing field: ${fieldName}");
 
+  fieldName = _getFieldNameFromDeclaration(variable);
   // check if there is a DartsonProperty annotation
-  SerializedName prop = variable.annotations?.firstWhere((a) => a is SerializedName, orElse: () => null);
-//  _serLog.finest("Property Annotation: ${prop}");
-
-  if (prop?.name != null) {
-//    _serLog.finer("Field renamed to: ${prop.name}");
-    fieldName = prop.name;
-  }
-
 
 //  _serLog.finer("depth: $depth");
 
   //If the value is not null and the annotation @ignore is not on variable declaration
-  if (value != null && !(variable.annotations?.any((a) => a is _Ignore) ?? false)
+  if (value != null && !_getIsIgnoredFromDeclaration(variable)
       // And exclude is pressent
       && (exclude == null
         // or exclude is Map (we are excluding nested attribute)