+++ test/deserializer/deserializer_generic_test.g.dart
@@ -183,11 +183,12 @@ abstract class _$MapClassSerializable extends SerializableMap {
 
 _SimpleClass__Constructor([positionalParams, namedParams]) => new SimpleClass();
 
-const $$SimpleClass_fields_name = const DeclarationMirror(type: String);
+const $$SimpleClass_fields_name =
+    const DeclarationMirror(name: 'name', type: String);
 
 const SimpleClassClassMirror =
     const ClassMirror(name: 'SimpleClass', constructors: const {
-  '': const FunctionMirror($call: _SimpleClass__Constructor)
+  '': const FunctionMirror(name: '', $call: _SimpleClass__Constructor)
 }, fields: const {
   'name': $$SimpleClass_fields_name
 }, getters: const [
@@ -203,12 +204,12 @@ const SimpleClassClassMirror =
 _ListClass__Constructor([positionalParams, namedParams]) => new ListClass();
 
 const $$ListClass_fields_list =
-    const DeclarationMirror(type: const [List, SimpleClass]);
+    const DeclarationMirror(name: 'list', type: const [List, SimpleClass]);
 
 const ListClassClassMirror = const ClassMirror(
     name: 'ListClass',
     constructors: const {
-      '': const FunctionMirror($call: _ListClass__Constructor)
+      '': const FunctionMirror(name: '', $call: _ListClass__Constructor)
     },
     fields: const {
       'list': $$ListClass_fields_list
@@ -222,7 +223,8 @@ const ListClassClassMirror = const ClassMirror(
 _ListListClass__Constructor([positionalParams, namedParams]) =>
     new ListListClass();
 
-const $$ListListClass_fields_list = const DeclarationMirror(type: const [
+const $$ListListClass_fields_list =
+    const DeclarationMirror(name: 'list', type: const [
   List,
   const [List, SimpleClass]
 ]);
@@ -230,7 +232,7 @@ const $$ListListClass_fields_list = const DeclarationMirror(type: const [
 const ListListClassClassMirror = const ClassMirror(
     name: 'ListListClass',
     constructors: const {
-      '': const FunctionMirror($call: _ListListClass__Constructor)
+      '': const FunctionMirror(name: '', $call: _ListListClass__Constructor)
     },
     fields: const {
       'list': $$ListListClass_fields_list
@@ -244,7 +246,8 @@ const ListListClassClassMirror = const ClassMirror(
 _ListListListClass__Constructor([positionalParams, namedParams]) =>
     new ListListListClass();
 
-const $$ListListListClass_fields_list = const DeclarationMirror(type: const [
+const $$ListListListClass_fields_list =
+    const DeclarationMirror(name: 'list', type: const [
   List,
   const [
     List,
@@ -255,7 +258,7 @@ const $$ListListListClass_fields_list = const DeclarationMirror(type: const [
 const ListListListClassClassMirror = const ClassMirror(
     name: 'ListListListClass',
     constructors: const {
-      '': const FunctionMirror($call: _ListListListClass__Constructor)
+      '': const FunctionMirror(name: '', $call: _ListListListClass__Constructor)
     },
     fields: const {
       'list': $$ListListListClass_fields_list
@@ -269,7 +272,8 @@ const ListListListClassClassMirror = const ClassMirror(
 _ListMapClass__Constructor([positionalParams, namedParams]) =>
     new ListMapClass();
 
-const $$ListMapClass_fields_list = const DeclarationMirror(type: const [
+const $$ListMapClass_fields_list =
+    const DeclarationMirror(name: 'list', type: const [
   List,
   const [
     Map,
@@ -280,7 +284,7 @@ const $$ListMapClass_fields_list = const DeclarationMirror(type: const [
 const ListMapClassClassMirror = const ClassMirror(
     name: 'ListMapClass',
     constructors: const {
-      '': const FunctionMirror($call: _ListMapClass__Constructor)
+      '': const FunctionMirror(name: '', $call: _ListMapClass__Constructor)
     },
     fields: const {
       'list': $$ListMapClass_fields_list
@@ -294,7 +298,8 @@ const ListMapClassClassMirror = const ClassMirror(
 _ListListMapClass__Constructor([positionalParams, namedParams]) =>
     new ListListMapClass();
 
-const $$ListListMapClass_fields_list = const DeclarationMirror(type: const [
+const $$ListListMapClass_fields_list =
+    const DeclarationMirror(name: 'list', type: const [
   List,
   const [
     List,
@@ -308,7 +313,7 @@ const $$ListListMapClass_fields_list = const DeclarationMirror(type: const [
 const ListListMapClassClassMirror = const ClassMirror(
     name: 'ListListMapClass',
     constructors: const {
-      '': const FunctionMirror($call: _ListListMapClass__Constructor)
+      '': const FunctionMirror(name: '', $call: _ListListMapClass__Constructor)
     },
     fields: const {
       'list': $$ListListMapClass_fields_list
@@ -321,7 +326,7 @@ const ListListMapClassClassMirror = const ClassMirror(
     ]);
 _MapClass__Constructor([positionalParams, namedParams]) => new MapClass();
 
-const $$MapClass_fields_map = const DeclarationMirror(type: const [
+const $$MapClass_fields_map = const DeclarationMirror(name: 'map', type: const [
   Map,
   const [String, SimpleClass]
 ]);
@@ -329,7 +334,7 @@ const $$MapClass_fields_map = const DeclarationMirror(type: const [
 const MapClassClassMirror = const ClassMirror(
     name: 'MapClass',
     constructors: const {
-      '': const FunctionMirror($call: _MapClass__Constructor)
+      '': const FunctionMirror(name: '', $call: _MapClass__Constructor)
     },
     fields: const {
       'map': $$MapClass_fields_map