+++ test/serializer/cyclic_reference_test.g.dart
@@ -296,15 +296,19 @@ abstract class _$CourseSerializable extends SerializableMap {
 
 _Employee__Constructor([positionalParams, namedParams]) => new Employee();
 
-const $$Employee_fields_id = const DeclarationMirror(type: int);
-const $$Employee_fields_firstName = const DeclarationMirror(type: String);
-const $$Employee_fields_lastName = const DeclarationMirror(type: String);
-const $$Employee_fields_address = const DeclarationMirror(type: Address);
-const $$Employee_fields_manager = const DeclarationMirror(type: Employee);
+const $$Employee_fields_id = const DeclarationMirror(name: 'id', type: int);
+const $$Employee_fields_firstName =
+    const DeclarationMirror(name: 'firstName', type: String);
+const $$Employee_fields_lastName =
+    const DeclarationMirror(name: 'lastName', type: String);
+const $$Employee_fields_address =
+    const DeclarationMirror(name: 'address', type: Address);
+const $$Employee_fields_manager =
+    const DeclarationMirror(name: 'manager', type: Employee);
 
 const EmployeeClassMirror =
     const ClassMirror(name: 'Employee', constructors: const {
-  '': const FunctionMirror($call: _Employee__Constructor)
+  '': const FunctionMirror(name: '', $call: _Employee__Constructor)
 }, annotations: const [
   cyclical
 ], fields: const {
@@ -328,16 +332,21 @@ const EmployeeClassMirror =
 ]);
 _Address__Constructor([positionalParams, namedParams]) => new Address();
 
-const $$Address_fields_id = const DeclarationMirror(type: int);
-const $$Address_fields_street = const DeclarationMirror(type: String);
-const $$Address_fields_city = const DeclarationMirror(type: String);
-const $$Address_fields_country = const DeclarationMirror(type: String);
-const $$Address_fields_postalCode = const DeclarationMirror(type: String);
-const $$Address_fields_owner = const DeclarationMirror(type: Employee);
+const $$Address_fields_id = const DeclarationMirror(name: 'id', type: int);
+const $$Address_fields_street =
+    const DeclarationMirror(name: 'street', type: String);
+const $$Address_fields_city =
+    const DeclarationMirror(name: 'city', type: String);
+const $$Address_fields_country =
+    const DeclarationMirror(name: 'country', type: String);
+const $$Address_fields_postalCode =
+    const DeclarationMirror(name: 'postalCode', type: String);
+const $$Address_fields_owner =
+    const DeclarationMirror(name: 'owner', type: Employee);
 
 const AddressClassMirror =
     const ClassMirror(name: 'Address', constructors: const {
-  '': const FunctionMirror($call: _Address__Constructor)
+  '': const FunctionMirror(name: '', $call: _Address__Constructor)
 }, annotations: const [
   cyclical
 ], fields: const {
@@ -364,14 +373,18 @@ const AddressClassMirror =
 ]);
 _Employee2__Constructor([positionalParams, namedParams]) => new Employee2();
 
-const $$Employee2_fields_firstName = const DeclarationMirror(type: String);
-const $$Employee2_fields_lastName = const DeclarationMirror(type: String);
-const $$Employee2_fields_address = const DeclarationMirror(type: Address2);
-const $$Employee2_fields_manager = const DeclarationMirror(type: Employee2);
+const $$Employee2_fields_firstName =
+    const DeclarationMirror(name: 'firstName', type: String);
+const $$Employee2_fields_lastName =
+    const DeclarationMirror(name: 'lastName', type: String);
+const $$Employee2_fields_address =
+    const DeclarationMirror(name: 'address', type: Address2);
+const $$Employee2_fields_manager =
+    const DeclarationMirror(name: 'manager', type: Employee2);
 
 const Employee2ClassMirror =
     const ClassMirror(name: 'Employee2', constructors: const {
-  '': const FunctionMirror($call: _Employee2__Constructor)
+  '': const FunctionMirror(name: '', $call: _Employee2__Constructor)
 }, annotations: const [
   cyclical
 ], fields: const {
@@ -392,15 +405,20 @@ const Employee2ClassMirror =
 ]);
 _Address2__Constructor([positionalParams, namedParams]) => new Address2();
 
-const $$Address2_fields_street = const DeclarationMirror(type: String);
-const $$Address2_fields_city = const DeclarationMirror(type: String);
-const $$Address2_fields_country = const DeclarationMirror(type: String);
-const $$Address2_fields_postalCode = const DeclarationMirror(type: String);
-const $$Address2_fields_owner = const DeclarationMirror(type: Employee2);
+const $$Address2_fields_street =
+    const DeclarationMirror(name: 'street', type: String);
+const $$Address2_fields_city =
+    const DeclarationMirror(name: 'city', type: String);
+const $$Address2_fields_country =
+    const DeclarationMirror(name: 'country', type: String);
+const $$Address2_fields_postalCode =
+    const DeclarationMirror(name: 'postalCode', type: String);
+const $$Address2_fields_owner =
+    const DeclarationMirror(name: 'owner', type: Employee2);
 
 const Address2ClassMirror =
     const ClassMirror(name: 'Address2', constructors: const {
-  '': const FunctionMirror($call: _Address2__Constructor)
+  '': const FunctionMirror(name: '', $call: _Address2__Constructor)
 }, annotations: const [
   cyclical
 ], fields: const {
@@ -424,14 +442,15 @@ const Address2ClassMirror =
 ]);
 _Student__Constructor([positionalParams, namedParams]) => new Student();
 
-const $$Student_fields_id = const DeclarationMirror(type: int);
-const $$Student_fields_name = const DeclarationMirror(type: String);
+const $$Student_fields_id = const DeclarationMirror(name: 'id', type: int);
+const $$Student_fields_name =
+    const DeclarationMirror(name: 'name', type: String);
 const $$Student_fields_courses =
-    const DeclarationMirror(type: const [List, Course]);
+    const DeclarationMirror(name: 'courses', type: const [List, Course]);
 
 const StudentClassMirror =
     const ClassMirror(name: 'Student', constructors: const {
-  '': const FunctionMirror($call: _Student__Constructor)
+  '': const FunctionMirror(name: '', $call: _Student__Constructor)
 }, annotations: const [
   cyclical
 ], fields: const {
@@ -449,14 +468,15 @@ const StudentClassMirror =
 ]);
 _Course__Constructor([positionalParams, namedParams]) => new Course();
 
-const $$Course_fields_id = const DeclarationMirror(type: int);
-const $$Course_fields_beginDate = const DeclarationMirror(type: DateTime);
+const $$Course_fields_id = const DeclarationMirror(name: 'id', type: int);
+const $$Course_fields_beginDate =
+    const DeclarationMirror(name: 'beginDate', type: DateTime);
 const $$Course_fields_students =
-    const DeclarationMirror(type: const [List, Student]);
+    const DeclarationMirror(name: 'students', type: const [List, Student]);
 
 const CourseClassMirror =
     const ClassMirror(name: 'Course', constructors: const {
-  '': const FunctionMirror($call: _Course__Constructor)
+  '': const FunctionMirror(name: '', $call: _Course__Constructor)
 }, annotations: const [
   cyclical
 ], fields: const {