+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 * [ ] Add include parameter to serialize function
 * [x] Add uuid check over hashCode for cyclical references
+* [ ] Add types to serialized values
+
+## 0.11.0
+
+- add caches for `SerializedName` and `Ignored` annotations
+- upgrade to `serializable@0.8.0` and `built_mirrors@0.7.1`
+
 ## 0.10.2
 
 - fixes #23: The getter 'name' was called on null